backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
nba_api>=1.4.1
python-dotenv>=1.0.1
supabase>=2.5.0
httpx>=0.27.0
pydantic>=2.7.4
matplotlib>=3.8.0
numpy>=1.26.0
scipy>=1.12.0
